+++ examples/gjs/gtk-3-browser/main.ts
@@ -58,7 +58,7 @@ wnd.connect('delete-event', () => {
     return true
 })
 entry.connect('activate', (self) => {
-    let uri = self.text
+    let uri = self.text || ''
     if (uri && !(uri.startsWith('http://') || uri.startsWith('https://') || uri.startsWith('ftp://'))) uri = 'http://' + uri
     webview.load_uri(uri)
 })